+++ src/cmd_pgr/mod.rs
@@ -9,6 +9,7 @@ pub mod fa;
 pub mod fas;
 pub mod fq;
 pub mod gff;
+pub mod kmer;
 pub mod lav;
 pub mod maf;
 pub mod ms;